Se cambia el componente de USER por LANDING y LOGIN separando la visualización de la página de login en dos partes, una con el componente de logueo y otra para el componente visual. El componente de logueo también se separa login en dos partes una con la parte visual del formulario LOGIN y otra para la función CONTROLLER
