--- src/User/User.css
@@ -1,3 +1,0 @@
-.login{
-    background-color: #E9478A;
-}
